third_party/blink/web_tests/printing/css2.1/page-break-after-002-expected.txt

This sentence must be on the first page.

This sentence must be on the second page.

This sentence must be either on the second or on the third page.


PASS: page number of "test1" is 1
PASS: page number of "test2" is 1
All tests passed
